+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add proposal review field
+- Add tab of related entities. [#524](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/524)
 
 ## [2.16.2](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/34) 2023-09-06